+++ src/utils/widgetUtils.ts
@@ -1,0 +1,16 @@
+export function formatWidth({
+  value,
+  unit,
+}: {
+  value: number;
+  unit: "px" | "%";
+}) {
+  switch (unit) {
+    case "px":
+      return `${value}px`;
+    case "%":
+      return `${value}%`;
+    default:
+      return `${value}px`;
+  }
+}